+++ src/main/java/com/chriscarini/jetbrains/logshipper/LogstashLayoutAppenderService.java
@@ -3,9 +3,11 @@ package com.chriscarini.jetbrains.logshipper;
 import com.chriscarini.jetbrains.logshipper.configuration.SettingsManager;
 import com.intellij.ide.AppLifecycleListener;
 import com.intellij.openapi.application.ApplicationManager;
+import com.intellij.openapi.application.ex.ApplicationInfoEx;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.util.ShutDownTracker;
 import com.intellij.openapi.util.text.StringUtil;
+import com.intellij.platform.ide.bootstrap.StartupUtil;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.IOException;
@@ -52,12 +54,17 @@ public class LogstashLayoutAppenderService implements AppLifecycleListener {
                 return;
             }
 
+            LOG.info("Adding Logshipper handler to root logger");
             // Add the handler to the root Logger
             getRootLogger().addHandler(this.handler);
 
             // Register a shutdown task to remove the handler and close it cleanly.
             // noinspection UnstableApiUsage
             ShutDownTracker.getInstance().registerShutdownTask(this::cleanupHandler);
+            
+            // Once we're attached to the root logger, log out the same essential information 
+            // about the IDE as normally happens during IDE startup.
+            StartupUtil.logEssentialInfoAboutIde(LOG, ApplicationInfoEx.getInstanceEx(), List.of(""));
 
             LOG.info("Added Logshipper handler to root logger");
         } else {
@@ -74,7 +81,9 @@ public class LogstashLayoutAppenderService implements AppLifecycleListener {
      */
     private void cleanupHandler() {
         if (this.handler != null) {
+            LOG.info("Removing Logshipper handler from root logger");
             getRootLogger().removeHandler(this.handler);
+            LOG.info("Closing Logshipper handler");
             this.handler.close();
             this.handler = null;
         }